fix unit test
