+++ engine/src/test/java/io/seldon/engine/api/rest/TestRestClientControllerExternalGraphs.java
@@ -10,6 +10,8 @@ import io.seldon.protos.DeploymentProtos.PredictorSpec;
 import io.seldon.protos.PredictionProtos.SeldonMessage;
 import java.net.URI;
 import java.nio.charset.StandardCharsets;
+
+import org.junit.After;
 import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
@@ -25,6 +27,7 @@ import org.springframework.boot.test.context.SpringBootTest.WebEnvironment;
 import org.springframework.boot.test.web.client.TestRestTemplate;
 import org.springframework.boot.web.server.LocalServerPort;
 import org.springframework.http.HttpEntity;
+import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
@@ -36,6 +39,8 @@ import org.springframework.test.web.servlet.MvcResult;
 import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
 import org.springframework.test.web.servlet.setup.MockMvcBuilders;
 import org.springframework.util.MultiValueMap;
+import org.springframework.web.client.HttpServerErrorException;
+import org.springframework.web.client.HttpStatusCodeException;
 import org.springframework.web.context.WebApplicationContext;
 
 @RunWith(SpringRunner.class)
@@ -57,6 +62,11 @@ public class TestRestClientControllerExternalGraphs {
   public void setup() throws Exception {
 
     mvc = MockMvcBuilders.webAppContextSetup(context).addFilters(new XSSFilter()).build();
+  }
+
+  @After
+  public void resetMocks() {
+    Mockito.reset(testRestTemplate.getRestTemplate());
   }
 
   @LocalServerPort private int port;
@@ -750,5 +760,41 @@ public class TestRestClientControllerExternalGraphs {
                 "mygauge2{deployment_name=\"None\",model_image=\"seldonio/mock_classifier\",model_name=\"mean-classifier\",model_version=\"0.6\",predictor_name=\"fx-market-predictor\",predictor_version=\"unknown\",} 100.0")
             > -1);
     System.out.println(response);
+  }
+
+  @Test
+  public void testModelPredictionNon200Response() throws Exception {
+    String jsonStr = readFile("src/test/resources/model_simple.json", StandardCharsets.UTF_8);
+    String responseStr =
+            readFile("src/test/resources/response_status.json", StandardCharsets.UTF_8);
+    io.seldon.protos.DeploymentProtos.PredictorSpec.Builder PredictorSpecBuilder = io.seldon.protos.DeploymentProtos.PredictorSpec.newBuilder();
+    EnginePredictor.updateMessageBuilderFromJson(PredictorSpecBuilder, jsonStr);
+    io.seldon.protos.DeploymentProtos.PredictorSpec predictorSpec = PredictorSpecBuilder.build();
+    final String predictJson = "{" + "\"binData\": \"MTIz\"" + "}";
+    ReflectionTestUtils.setField(enginePredictor, "predictorSpec", predictorSpec);
+
+    HttpStatusCodeException exception = HttpServerErrorException.InternalServerError
+            .create(HttpStatus.BAD_REQUEST, "status text", HttpHeaders.EMPTY, responseStr.getBytes(StandardCharsets.UTF_8), StandardCharsets.UTF_8);
+
+    Mockito.when(
+            testRestTemplate
+                    .getRestTemplate()
+                    .postForEntity(
+                            Matchers.<URI>any(),
+                            Matchers.<HttpEntity<MultiValueMap<String, String>>>any(),
+                            Matchers.<Class<String>>any()))
+            .thenThrow(exception);
+
+    MvcResult res =
+            mvc.perform(
+                    MockMvcRequestBuilders.post("/api/v0.1/predictions")
+                            .accept(MediaType.APPLICATION_JSON_UTF8)
+                            .content(predictJson)
+                            .contentType(MediaType.APPLICATION_JSON_UTF8))
+                    .andReturn();
+
+    // Check for returned response that wraps the ApiException into SeldonMessage
+    Assert.assertEquals(200, res.getResponse().getStatus());
+    Assert.assertEquals(responseStr, res.getResponse().getContentAsString());
   }
 }
